Testing and bug fixing

+++ app/controllers/calculations_controller.rb
@@ -9,7 +9,13 @@ class CalculationsController < ApplicationController
     end
 
     def edit
-        @calculation = Calculation.find(params[:calculation_id])
+        @user = current_user
+        if @user.calculations.exists?(params[:calculation_id])
+            @calculation = @user.calculations.find(params[:calculation_id])                                                    
+        else
+            flash[:alert] = "You do not have this calculation!"
+            redirect_to root_path
+        end
     end
 
     def update
@@ -36,7 +42,7 @@ class CalculationsController < ApplicationController
 
     def show
         @user = current_user
-        if @user.calculations.find_by(id: params['calculation_id'])
+        if @user.calculations.exists?(id: params['calculation_id'])
             @calculation = @user.calculations.find_by(id: params['calculation_id'])
             @top_three = @calculation.calculation_records.group_by { |r| r["total"] }
                                                             .sort_by  { |k, v| -k }
@@ -51,29 +57,33 @@ class CalculationsController < ApplicationController
     def create
         params = calculation_params
         params['user'] = current_user
-        infoFromDb = ApiService.getRates(params["base_currency"], params["target_currency"])
-        if infoFromDb != "ERROR"
-            todaysRate = infoFromDb[0]
-            infoFromDb.delete_at(0)
-            params['date'] = Date.parse(todaysRate['date'])
-            params['rate'] = todaysRate['rates']
-            @calculation = Calculation.new(params)
-            if @calculation.save()
-                if CalculationRecord.SaveRecordsForCalculation(todaysRate, infoFromDb, params['base_amount'].to_f, @calculation)
-                    redirect_to calculations_path
-                    else
+        if params['base_currency'] == params['target_currency']
+            flash[:alert] = "Base And Target currencies cannot be the same"
+            redirect_to calculation_new_path
+        else
+            infoFromDb = ApiService.getRates(params["base_currency"], params["target_currency"])
+            if infoFromDb != "ERROR"
+                todaysRate = infoFromDb[0]
+                infoFromDb.delete_at(0)
+                params['date'] = Date.parse(todaysRate['date'])
+                params['rate'] = todaysRate['rates']
+                @calculation = Calculation.new(params)
+                if @calculation.save()
+                    if CalculationRecord.SaveRecordsForCalculation(todaysRate, infoFromDb, params['base_amount'].to_f, @calculation)
+                        redirect_to calculations_path
+                        else
+                        flash[:alert] = "There was a problem with your calculation, please try again. Cannot save Calculation to Database."
+                        redirect_to calculation_new_path
+                    end
+                else
                     flash[:alert] = "There was a problem with your calculation, please try again. Cannot save Calculation to Database."
                     redirect_to calculation_new_path
                 end
             else
-                flash[:alert] = "There was a problem with your calculation, please try again. Cannot save Calculation to Database."
+                flash[:alert] = "There was a problem with your calculation, please try again. API responded with error"
                 redirect_to calculation_new_path
             end
-        else
-            flash[:alert] = "There was a problem with your calculation, please try again. API responded with error"
-            redirect_to calculation_new_path
         end
-
     end
